Further simplified creation extensibility
- An implementor of pipeline implements instance of BluePipelineCreateRequest and handles creation.
- Reuqest must have $class element set to the fully qualified class name of concrete instance of BluePipelineCreateRequest
