--- blueocean-github-pipeline/src/main/java/io/jenkins/blueocean/blueocean_github_pipeline/GithubPipelineCreatorImpl.java
@@ -1,121 +1,0 @@
-package io.jenkins.blueocean.blueocean_github_pipeline;
-
-import com.cloudbees.plugins.credentials.common.StandardCredentials;
-import hudson.Extension;
-import hudson.model.Cause;
-import hudson.model.TopLevelItem;
-import io.jenkins.blueocean.commons.ServiceException;
-import io.jenkins.blueocean.rest.Reachable;
-import io.jenkins.blueocean.rest.impl.pipeline.OrganizationFolderPipelineImpl;
-import io.jenkins.blueocean.rest.model.BluePipeline;
-import io.jenkins.blueocean.rest.model.BluePipelineCreateRequest;
-import io.jenkins.blueocean.rest.model.BluePipelineCreator;
-import io.jenkins.blueocean.rest.model.BlueScmConfig;
-import io.jenkins.blueocean.service.embedded.rest.AbstractPipelineCreateRequestImpl;
-import jenkins.branch.CustomOrganizationFolderDescriptor;
-import jenkins.branch.OrganizationFolder;
-import jenkins.model.Jenkins;
-import org.jenkinsci.plugins.github_branch_source.Connector;
-import org.jenkinsci.plugins.github_branch_source.GitHubSCMNavigator;
-
-import javax.annotation.Nonnull;
-import java.io.IOException;
-import java.util.List;
-
-/**
- * @author Vivek Pandey
- */
-@Extension
-public class GithubPipelineCreatorImpl extends BluePipelineCreator {
-    private static final String CREATOR_ID = GithubPipelineCreatorImpl.class.getName();
-
-    @Override
-    public String getId() {
-        return CREATOR_ID;
-    }
-
-    @Override
-    public Class<? extends BluePipelineCreateRequest> getType() {
-        return RequestImpl.class;
-    }
-
-    public static class RequestImpl extends AbstractPipelineCreateRequestImpl{
-        private static final String DESCRIPTOR = "jenkins.branch.OrganizationFolder.org.jenkinsci.plugins.github_branch_source.GitHubSCMNavigator";
-        private String name;
-
-        private BlueScmConfig scmConfig = new BlueScmConfig();
-
-        @Nonnull
-        @Override
-        public String getName() {
-            return name;
-        }
-
-        @Override
-        public void setName(String name) {
-            this.name = name;
-        }
-
-        public BlueScmConfig getScmConfig() {
-            return scmConfig;
-        }
-
-        public void setScmConfig(BlueScmConfig scmConfig) {
-            this.scmConfig = scmConfig;
-        }
-
-        @SuppressWarnings("unchecked")
-        @Override
-        public BluePipeline create(Reachable parent) throws IOException {
-
-            String apiUrl=null;
-            String orgName = name; //default
-            String credentialId = null;
-            StringBuilder sb = new StringBuilder();
-
-            if(scmConfig != null) {
-                apiUrl = scmConfig.getUri();
-                if(scmConfig.getConfig().get("orgName") instanceof String){
-                    orgName = (String) scmConfig.getConfig().get("orgName");
-                }
-                credentialId = scmConfig.getCredentialId();
-                if(scmConfig!= null && scmConfig.getConfig().get("repos") instanceof List) {
-                    for (String r : (List<String>)scmConfig.getConfig().get("repos")) {
-                        sb.append(String.format("(%s\\b)?", r));
-                    }
-                }
-            }
-            TopLevelItem item = create(Jenkins.getInstance(),name, DESCRIPTOR, CustomOrganizationFolderDescriptor.class);
-
-            if(item instanceof OrganizationFolder){
-                GitHubSCMNavigator gitHubSCMNavigator = new GitHubSCMNavigator(apiUrl, orgName, credentialId, credentialId);
-                if(sb.length() > 0){
-                    gitHubSCMNavigator.setPattern(sb.toString());
-                }
-                validateCredentialId(credentialId, (OrganizationFolder) item, gitHubSCMNavigator);
-
-                // cick of github scan build
-                OrganizationFolder organizationFolder = (OrganizationFolder) item;
-                organizationFolder.getNavigators().replace(gitHubSCMNavigator);
-                organizationFolder.scheduleBuild(new Cause.UserIdCause());
-                return new OrganizationFolderPipelineImpl(organizationFolder, parent.getLink());
-            }
-            return null;
-        }
-
-        private void validateCredentialId(String credentialId, OrganizationFolder item, GitHubSCMNavigator navigator) throws IOException {
-            if(credentialId != null && !credentialId.trim().isEmpty()) {
-                StandardCredentials credentials = Connector.lookupScanCredentials(item, navigator.getApiUri(), credentialId);
-                if (credentials == null) {
-                    try {
-                        item.delete();
-                    } catch (InterruptedException e) {
-                        throw new ServiceException.UnexpectedErrorException("Invalid credentialId: " + credentialId + ". Failure during cleaing up folder: " + item.getName() + ". Error: " + e.getMessage(), e);
-                    }
-                    throw new ServiceException.BadRequestExpception("Invalid credentialId: " + credentialId);
-                }
-            }
-        }
-    }
-
-}
