+++ blueocean-rest/src/main/java/io/jenkins/blueocean/rest/model/BluePipelineContainer.java
@@ -1,8 +1,10 @@
 package io.jenkins.blueocean.rest.model;
 
 import io.jenkins.blueocean.commons.ServiceException;
-import io.jenkins.blueocean.commons.stapler.JsonBody;
+import net.sf.json.JSONObject;
+import org.kohsuke.stapler.StaplerRequest;
 import org.kohsuke.stapler.WebMethod;
+import org.kohsuke.stapler.json.JsonBody;
 import org.kohsuke.stapler.verb.POST;
 
 import java.io.IOException;
@@ -17,12 +19,20 @@ public abstract class BluePipelineContainer extends Container<BluePipeline>{
     /**
      * Create new pipeline.
      *
-     * @param request {@link BluePipelineCreateRequest} request object
+     * @param body {@link BluePipelineCreateRequest} request object
      * @return {@link CreateResponse} response
      */
     @POST
     @WebMethod(name = "")
-    public  CreateResponse create(@JsonBody BluePipelineCreateRequest request) throws IOException{
+    public  CreateResponse create(@JsonBody JSONObject body, StaplerRequest staplerRequest) throws IOException{
+        if(body.get("$class") == null){
+            throw new ServiceException.BadRequestExpception("$class is required element");
+        }
+        BluePipelineCreateRequest request = staplerRequest.bindJSON(BluePipelineCreateRequest.class, body);
+        return create(request);
+    }
+
+    public CreateResponse create(BluePipelineCreateRequest request) throws IOException {
         if(request.getName() == null){
             throw new ServiceException.BadRequestExpception("name is required element");
         }
