+++ blueocean-rest/src/main/java/io/jenkins/blueocean/rest/model/BlueScmConfig.java
@@ -1,5 +1,8 @@
 package io.jenkins.blueocean.rest.model;
 
+import net.sf.json.JSONObject;
+import org.kohsuke.stapler.DataBoundConstructor;
+
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 import java.util.LinkedHashMap;
@@ -16,6 +19,15 @@ public class BlueScmConfig {
 
     private Map<String, Object> config = new LinkedHashMap<>();
 
+    @DataBoundConstructor
+    public BlueScmConfig(String uri, String credentialId, JSONObject config) {
+        this.uri = uri;
+        this.credentialId = credentialId;
+        this.config = config;
+    }
+
+
+
     /**
      * Gives URI of SCM
      */
@@ -23,9 +35,9 @@ public class BlueScmConfig {
         return uri;
     }
 
-    public void setUri(@Nullable String uri) {
-        this.uri = uri;
-    }
+//    public void setUri(@Nullable String uri) {
+//        this.uri = uri;
+//    }
 
     /**
      * Gives credential id
@@ -34,9 +46,9 @@ public class BlueScmConfig {
         return credentialId;
     }
 
-    public void setCredentialId(String credentialId) {
-        this.credentialId = credentialId;
-    }
+//    public void setCredentialId(String credentialId) {
+//        this.credentialId = credentialId;
+//    }
 
     /**
      * Gives map of configuration item specific to this pipeline creation
@@ -45,7 +57,7 @@ public class BlueScmConfig {
         return config;
     }
 
-    public void setConfig(@Nonnull Map<String, Object> config) {
-        this.config.putAll(config);;
-    }
+//    public void setConfig(@Nonnull Map<String, Object> config) {
+//        this.config.putAll(config);;
+//    }
 }
